Externalized ValidateScope to a seperate file within the auth package
