+++ auth/auth.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/auth0/go-jwt-middleware/v2/jwks"
 	"github.com/auth0/go-jwt-middleware/v2/validator"
-	"github.com/gin-gonic/gin"
 	"github.com/spf13/viper"
 )
 
@@ -49,19 +48,4 @@ func GetTokenValidator() (*validator.Validator, error) {
 	}
 
 	return tokenValidator, nil
-}
-
-/*
-ValidateScope Fetch validated claims from the gin context and ensure that
-the user has the desired scope
-*/
-func ValidateScope(ctx *gin.Context, requiredScope string) bool {
-	token := ctx.Value("token").(*validator.ValidatedClaims)
-
-	claims := token.CustomClaims.(*CustomClaims)
-	if !claims.HasScope(requiredScope) {
-		return false
-	}
-
-	return true
 }
